requirements-docs.txt
reportlab
python-docx
openpyxl
pypdf
